Stable Version Update
It provides frame per second (FPS) function and Pyplot (plt) styling features. Also, raspberry pi is supported in this version, added two parameters in a function to prevent memory leak. Besides, this version includes a 39 second music for your trial. Run it on Jupyter Notebook is recommended. Finally, if you facing problem or error, please feel free to leave a comment in issues of this repository.